+++ AddictionBreaker/app/src/main/java/com/example/addictionbreaker/data/DatabaseHelper.java
@@ -1,4 +1,4 @@
-package com.example.addictionbreaker;
+package com.example.addictionbreaker.data;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
