+++ AddictionBreaker/app/src/main/res/layout/activity_home.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".ui.HomeActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
